requirements.txt
requests>=2.31.0
schedule>=1.2.0
facebook-business>=17.0.0
flask>=2.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.11.0
smtplib-ssl>=1.0.0
email-validator>=2.0.0
cryptography>=41.0.0
python-dateutil>=2.8.0
pytz>=2023.3